[TM-874] download site geojson
